Use the new deployments orchestrator by default (#1922)
If you need the 'classic' orchestrator, then the UI can be enabled in
`runtime.exs`, and the deployment can be switched over, but the new
orchestrator should be the default.

At NervesCloud, we will be using the new orchestrator and not enabling
the UI to switch to the old one.