Managed Deployments: Groups and Releases (#1653)
This work contains two sets of changes: the creation of
`DeploymentRelease`s and the renaming of `deployments` to
`deployment_groups`. `DeploymentRelease` integration will be done in the
next body of work for managed deploys, this PR has become too big to add
any more functionality.

So many things have been renamed that it's easier to list what _hasn't_
been renamed:
- `deployments` table name
- any indexes with `deployments` in the name
- `deployment_id` foreign key in DB and application code
- pub sub topic and events

Database-related renaming will be done later. Pub sub topic and events
should and will be adjusted, but this is not the time. It won't be a
simple rename and needs more care and thought.