Cache the `product_id` in the `DeviceConnection` (#1973)
This is required for improved query perf when polling for the count of
connected devices grouped by `product_id`.

The way to do such a query, before these changes, required a join, and
that itself just added more for the query planner to take care of.

This work is a required part of an Insights dashboard I'm working on.

---------

Co-authored-by: Nate Shoemaker <transition32@gmail.com>