Make sure 'skip queue' firmware update message are sent to devices (#1985)
I introduced some poor logic which impacted 'skip queue' calls.

The message was only sent to devices if they were using the new
distributed orchestrator.

This fixes the issue by leaving the message sending to `told_to_update`,
and making sure the pid of the device channel is passed through if the
old message passing pattern is required.

Oh, also, the message that was being sent wasn't actually doing
anything, so this cleans that up.