Add proper assigns to DeviceDetailsPage, ensure firmware is loaded on deployment group (#2009)
This is a fix for two issues we're seeing in our most recent release.

1.) The `firmware` association needs to be loaded before work in
`Device.told_to_update/3`. Jon pointed out we're preloading during
tests, so this was easily missed. I updated the tests to _not_ preload
firmware.
2.) `user` and `org_user` assigns were missing from the
DeviceDetailsPage socket. The cause was
https://github.com/nerves-hub/nerves_hub_web/commit/6b423f59be1b2a6e56c66246e23e210a6bed519f#diff-fe77162690c4853d5e459262184c9ada52018dac27c388bde3e97f2bc6dad980L37,
where the blanket `assign(assigns)` was removed.