Support device index table row links in webkit (#2025)
Webkit doesn't support `display: relative` on `<tr />`'s. That means
that our absolute-positioned table row mask, which allows the entire
table row to be clickable, takes up the entire page 😖 Using a mix of
`clip-path` and `transform: translate()` is enough to tell Webkit to
keep the bounds in the table row.

#### Webkit pre-fix
![CleanShot 2025-04-05 at 07 35
56](https://github.com/user-attachments/assets/525e41bb-79c0-4714-be64-020ceca61c1f)

#### Webkit post-fix
![CleanShot 2025-04-05 at 07 33
06](https://github.com/user-attachments/assets/3d631de5-30d3-4522-b3ba-26f86537e144)

#### Chromium post-fix
![CleanShot 2025-04-05 at 07 31
26](https://github.com/user-attachments/assets/1942e538-315b-4aa0-ae17-1da4661e5647)