Ensure generated user token is Base62 safe
Base62 uses integer based calculations since 62 has no shared exponents
with 256 and therefor cannot be bit aligned. Because of this, leading
zeros/null bytes in a binary get ignored since Erlang is big-endian by
default and `0123` is the same as `123`.

If we attempt to add `0` padding to an encoding, then we potentially
fail decoding since we don't know if that incoming encoded string is
supposed to be `"0"` or `<<0>>`.

Therefor, to continue using Base62, this adds a function to loop through
token creation until a non-zero-starting token is made. In testing, this
was ~1/100 times, so it should be okay for a little extra work since
user tokens are also not generated frequently