Use hooks instead of LiveComponents (#2026)
When I built the tabs for the Device page, and then used a similar
pattern for the Deployment page, I quickly learnt that LiveComponents
have a lot of quicks which lead to confusing code and duplicated state
between the parent and child component.

For example, LiveComponents can't subscribe to topic, listen for pubsub
messages, or respond to timer events.

I recently discovered a better path forward which uses LiveView hooks :
https://hexdocs.pm/phoenix_live_view/Phoenix.LiveView.html#attach_hook/4

I've created some reusable code for implementing the same pattern with
the Deployment page, and the upcoming Org pages.