A revamp of our user token storage and management (#2024)
Heavily inspired by the Phoenix auth generators, including the API token
guide. This improves many aspects of how we store and use tokens, as
well as generating session tokens so we can delete them if a user's
account is compromised.

Key improvements:

- session and api tokens are entirely random and don't include any
components which can be guessed
- store session tokens, allowing for sessions to be cancelled/deleted by
a user or admin
- add a 'context' to the user token so we can track different types of
tokens issues

While there are some functions which I have included but aren't used,
these will be used in the coming weeks as I update the login UI.

---------

Co-authored-by: Jon Carstens <jon.carstens@smartrent.com>
Co-authored-by: Nate Shoemaker <transition32@gmail.com>