Refactor `UserToken` to reduce required code calls
More calls on the oncoming token binary were introduced in order to support
the old token requests as well. However, this can be simplified with a
pattern match since the old token strings are guaranteed to be 36 bytes
and the new friendly token strings will be 48+

This prevents extra calls by dictating which code path based on the
pattern match for string size. It also makes it a little easier for
cleanup in the future when V1 is completely deprecated and removed