OrgUser API improvements (#2063)
I saw a Sentry error pop through the NervesCloud account late last week
which was caused by the user trying to add a user to an org via the CLI.

The issue was that the CLI was asking for a username, but the API
expects a users ID.

We don't expose the ID of the users in other related API/CLI calls, plus
IDs aren't really friendly.

This PR:

- favors a users email for adding, removing, and updating their org
membership
- added a new `/invite` endpoint
- and complete OpenAPI specs

This is a breaking change. In my opinion, I think the change is fine to
merge, although thats just my opinion.

I have a CLI PR in the works which will use these updated routes.

---------

Co-authored-by: Nate Shoemaker <nshoes@users.noreply.github.com>