Monitor deployment processes in worker and report to Sentry (#2085)
Every 15min we check the counts of how many should be running and how
many are running. If the count is off we create a Sentry report with the
counts. Alongside putting together a timeline from logs I think this
will aid in debugging why our deployment orchestration processes aren't
starting properly on deploys.