Don't require tags when updating deployment group, changeset refactor (#2107)
This started as an adjustment to deployment group tag validations and
ended up being a refactor of deployment group changesets. The core issue
is that we handled tag validation differently in the create and update
changesets.

When updating a deployment group, tags are required, even though empty
tags is a valid state; you can create a deployment group with no tags.
This requirement has been removed. In application code, we treat
`%{"version" => ""}` as the empty state for `version`. Nothing has
stopped us in the changeset or database layer from saving `conditions`
to anything but a map with those two keys. Luckily, in SR prod, there
are no deployment group rows with one of these keys; it's always both. I
added the ability to partially update conditions so you don't have to
always provide both keys. This isn't used in application code but is in
tests and is a quality-of-life update.

Shared validations are now reconciled into a base changeset. The create
and update changesets then handle their own unique validations. This is
all tested and laid out nicely in
`NervesHub.ManagedDeployments.DeploymentGroupTest`. I tried a new
approach to organizing the tests, looking for feedback.

I then removed the ancient, recursive spaghetti code that would
magically add the `product_id` when creating a deployment group. This
has been simplified to requiring the `product_id` to be passed in
explicitly. This change avoids a database query when preloading firmware
to get `product_id`. The code in question handled a lot of different use
cases but preloading was the only one application code used :( In all
use cases, the firmware was already loaded making it trivial to pass in
`firmware.product_id`.

I also caught a route mistake when creating deployment groups in the new
UI.

Some changes I made were already covered by tests; I added tests for the
rest.

I had to update a two-year-old script in `priv/repo`, but I'm thinking
the whole thing can be removed, as well as the other script in the same
directory. Fine to pull the trigger if someone else double-checks me.