Increase DB timeout when mass moving devices in and out of deployment groups (#2130)
We use `Repo.update_all` to move and remove mass devices from deployment
groups. These will timeout given enough devices, around ~80,000. I had a
few different plans on how to batch this work, but it's simpler to let
the DB do the work over time. This uses a single connection, and the
feature isn't used often, so backing up the database isn't probable.

Looking for reasons why this is a bad idea :)