Fix metrics time frame not updating when clicking buttons (#2146)
Fixes #2104 where time frame was not updating when clicking buttons in
device health tab, because `
@default_time_frame` was always assigned on `push_patch`.

When adding tests for this I found out that charts did not update when
health is reported.
I don't know if that was intentional for some reason, I'll revert the
fix included in this PR if so.