Add information about available deltas to deployment summary (#2139)
This should let users see that deltas exist.

Also included the size of the delta file as well as the size of the full
target firmware. Until we collect stats on how much transfer we save
this is at least an indication.

It only shows up for products with deltas enabled. Older generated
deltas may not have the size information so they won't show it.