Safer delta updates
Changes to make delta updates a lot safer for anyone running older versions of fwup.

Adds detailed feature and version checking for deltas, FAT-support and encryption
Adds storage of update tool (fwup) metadata
Rework the plumbing of when deltas and full updates get sent
NervesHub now uses the Confuse library to pull information about what features of fwup are used and Confuse knows what fwup versions are needed for those features. Essentially we needed to know "given the fwup version on this device currently, can it support a delta update?" we can also figure out if there could be a problem doing a full update but that's not particularly in scope here.

This means fwup minimum version numbers mostly show up in Confuse rather than NervesHub.

As we want to save the information from Confuse while we have the firmware file ready for upload I added tool_metadata to both Firmware and FirmwareDelta schemas. I did not call it fwup_metadata as there are plans to add support for other tools and since I was touching this code I might as well save us some work for the future.

The DeltaUpdater grew some more responsibilities, such as figuring out if a delta update is allowed for this particular tool (fwup) and it is now named UpdateTool to reflect the wider responsibilites. Some callbacks were added for figuring out when delta or full update is appropriate. Most of the callbacks got updated signatures to allow more data to be sent back for storage. We had a messy way of doing this before to smuggle some metadata into upload_metadata I believe. This makes most of those fields on Firmware and FirmwareDelta.

Some of this metadata would be very nice to backfill. That is not covered in this work but we should have safe defaults and we can do that in a separate PR.

Since the on/off switch for deltas is already moved to the DeploymentGroup we have a way to control Delta vs Full updating of many devices in bulk.

This also adds a button to allow choosing to send a delta or a full update from the device details view.

Unzip is added as a dependency to perform some unpacking of meta.conf. I expect we'll use it more in the future to slurp meta.conf from uploaded archives without needing to download the whole thing.