Fix device status updates while viewing device index (#2178)
When we moved to a LiveView async result for device statuses, the
`handle_info`s for `connection:status` and `connection:update`
broadcasts weren't updated. This caused them to no-op and never update
the online status.

This work checks the proper key on `socket.assigns.device_statuses`,
updates the state with a proper `AsyncResult`, and adds a test for each
event.