Fix firmware uploads from API
The recent change to `DynamicConfigMultipart` assumed Application config
values were available during the `c:init/1` callback. However, plugs by
default init at compile time. So when deployed to production, this would
evaluate a `max_firmware_size` value of `nil` which would result in errors.

Also, the `DynamicConfigMultipart` test that was added completely bypassed
this type of check since it was manually initializing the module and calling
plug functions directly to verify behavior.

This changes to read the config value at runtime when a multipart upload for
a firmware file is attempted. It also removes the unneeded test module and
instead adds an integration level test to the firmware controller test to
assert the upload succeeds or is rejected for being too large.