Fix firmware uploads from API (#2198)
The recent change to `DynamicConfigMultipart` assumed Application config
values were available during the `c:init/1` callback. However, plugs by
default init at compile time. So when deployed to production, this would
evaluate a `max_firmware_size` value of `nil` which would result in
errors.

```
ts=2025-07-04T16:26:37.181 level=error msg="** (RuntimeError) :length option for Plug.Parsers.MULTIPART must be an integer
    (plug 1.18.0) lib/plug/parsers/multipart.ex:109: Plug.Parsers.MULTIPART.init/1
    (nerves_hub 2.0.0+6553d935) iex:49: NervesHubWeb.DynamicConfigMultipart.parse/5
    (plug 1.18.0) lib/plug/parsers.ex:341: Plug.Parsers.reduce/8
    (nerves_hub 2.0.0+6553d935) lib/nerves_hub_web/endpoint.ex:1: NervesHubWeb.Endpoint.plug_builder_call/2
    (nerves_hub 2.0.0+6553d935) lib/nerves_hub_web/endpoint.ex:1: NervesHubWeb.Endpoint.\"call (overridable 3)\"/2
    (nerves_hub 2.0.0+6553d935) lib/nerves_hub_web/endpoint.ex:1: NervesHubWeb.Endpoint.call/2
    (bandit 1.7.0) lib/bandit/pipeline.ex:131: Bandit.Pipeline.call_plug!/2
    (bandit 1.7.0) lib/bandit/pipeline.ex:42: Bandit.Pipeline.run/5
```

Also, the `DynamicConfigMultipart` test that was added completely
bypassed this type of check since it was manually initializing the
module and calling plug functions directly to verify behavior.

This changes to read the config value at runtime when a multipart upload
for a firmware file is attempted. It also removes the unneeded test
module and instead adds an integration level test to the firmware
controller test to assert the upload succeeds or is rejected for being
too large.