Firmware UUIDs are unique to products (#2233)
This should address #2232 

The main fix for this is scoping the `Firmwares.get_firmware_by_uuid` to
include the `product_id` by switching to
`Firmwares.get_firmware_by_product_and_uuid`.

I also tweaked the `with` matchings and the log messages to address a
Dialyzer warning after my initial change.