Handle `ArgumentError`s from `Plug.Crypto.non_executable_binary_to_term` (#2269)
This was found in the NervesCloud Sentry.

If an invalid device identifier, this this case an Erlang term, is
encoded in the shared secret auth,
`Plug.Crypto.non_executable_binary_to_term/2` correctly catches it, but
an error is raised which we don't handle nicely.

This PR rescues `ArgumentError`s, logs a message, and rejects the auth
request.