Add checks in delta creation to avoid creating larger deltas
If the overall delta firmware archive ends up larger than the target
delta firmware was. The delta generation process is simply failed. This
will let the system use the full target firmware instead which is more
efficient.

Per file to patch we now check that it is at least large enough to beat
the overhead of a delta patch (22 bytes). We also check that the
resulting patch file is smaller, otherwise we use the original file.
fwup has the smarts to figure out if this is done, so this is safe.