Update the Phoenix Socket topic mappings (fixes a fastlaning issue) (#2325)
The Phoenix Socket routes messages between connected clients, and as
such, it maintains a map of topics and PIDs to facilitate routing.

With our recent update, which changed the underlying Device subscription
from `device` to `device:[id]`, we need to update the mappings held in
the Socket state and ensure that the device ID is added to the topic.

While I'm not super in love with the breadth of these changes, this is
required, and I will return to this code and see if there are cleaner
options.