Move device connection pubsub updates to the `Connections` context (#2336)
While the Tracker gets a rewrite, this improves a bit of code I came
across where we announce a device online twice, once in the
`DeviceSocket` and once in `DeviceChannel`.

I still send two, but now I send `connecting` and `online`.

I've also encapsulated all connection related pubsub messages to
`Tracker`, and the `Connections` context is the primary user of Tracker.