Add a firmware validation filter when checking which devices are available for updating (#2369)
If you have a device that has just come online and is yet to validate
its firmware, and then another firmware is assigned to the deployment,
its possible for that device to be sent the update even though Fwup will
reject the update attempt until the firmware is validated.

Because Link doesn't send these errors to hub, the InflightUpdate is
left in limbo, causing the device to not be sent the update for an hour,
when the InflightUpdate is cleared away.

For devices that aren't sending firmware validation information to Hub
(older Link versions), the filter allows for devices we don't know the
validation status of, which just keeps the status quo logic.

Co-authored-by: Nate Shoemaker <nshoes@users.noreply.github.com>