Make devices index view much faster by deferring filter loading
Filter loading is now done using async. Filters are accessed via a
button and slide-out pane. We don't need the values for the dropdown
menus fully populated at the cost of rendering the device list.

This should be a significant improvement since we've seen the alarms
query be quite slow in production. That should also be addressed of
course but this removes a lot of work from the critical path of this
important view.