Use scripts plumbing for connecting code (#2399)
Solves #2395. 

This uses `scripts/run` to push connecting code to the device rather
than broadcasting it to every device node.

I added telemetry events for when the connecting code succeeds or fails,
as well as logging when it fails. The contract for "failing" a script
run is in NHL, which I don't love, but can be fixed at a later time.

---------

Co-authored-by: Jon Carstens <jon.carstens@smartrent.com>
Co-authored-by: Jon Carstens <jjcarstens@me.com>