Add devices online/offline to org and product views (#2402)
This is primarily to help people doing development or with few devices
but a couple of orgs set up. It helps for figuring out where your active
devices actually are. I've had this issue a hilarious number of times.

The loading is deferred so it won't stop someone from navigating if it
becomes slow. There are only two queries made so it shouldn't be
particularly impactful in performance and the information isn't
live-updated.

It might also indicate when something is really wrong or weird I
suppose. But mostly it helps indicate which orgs and products are
actually in use. Especially since the development seed produces a bunch
of them.

---------

Co-authored-by: Josh Kalderimis <josh.kalderimis@gmail.com>