Make device list refresh in a live but safe way (#2260)
For the purpose of making it possible to watch live updates more
information should refresh on connection changes. But it should be
throttled in a responsible way.

We can even let fwup_progress events indicate that things should refresh
if the pace is safe.

One big addition to making this safe is to determine which tabs are not
being watched and if they are not watche we don't update them.

This improves on the 10 second timer updater.

---------

Co-authored-by: Nate Shoemaker <transition32@gmail.com>