Introduce a priority queue (#2413)
This PR introduces priority queues to deployment groups.

The problem:

1. Often times you'll have a giant batch of hardware sitting in a
warehouse that is already preloaded with some old firmware version.
2. Installers then go into the field and set up these devices, but they
need to update right away since they could have been sitting in the
warehouse for ages, with some known to be buggy version, or a version
that might not even work with current APIs the hardware might need to
reach out to.
3. Installers then have to wait around while the device updates, often a
very expensive wait at that.

Currently the update could be clogged up behind the standard update
pool, perhaps behind a bunch of devices taking up all the slots due to
slow connections. This problem only compounds if installers are
installing multiple devices. What we really want is for those devices to
be updated independent of the in-use devices, and without having to go
through the process of making an entire duplicated deployment group for
them.

Solution:

Set a threshold version for firmware, under which any device instead
enters a priority queue that has its own size configurations. This
allows newly installed devices to immediately get updates without having
to increase the size of the normal queue, which would inadvertently
affect devices that are already out on the field.

<img width="1253" height="828" alt="Screenshot 2025-12-10 at 11 18
21 PM"
src="https://github.com/user-attachments/assets/80694f4e-ffc1-4205-8c1b-2b083098196b"
/>